refactor reader
